TST: Add additional progress tests
